pyproject.toml
[project]
name = "dotbackup"
dynamic = ["version"]
authors = [
	{ name="Jax Young", email="jaxvanyang@gmail.com" },
]
description = """\
	YAML config based backup utility. Easy to use yet flexible. With a primary focus on \
	dotfile backup & setup, but not limited to dotfiles.\
	"""
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["ruamel.yaml"]
keywords = ["YAML", "dotfiles", "backup"]
classifiers = [
	"Programming Language :: Python :: 3",
	"License :: OSI Approved :: MIT License",
	"Operating System :: POSIX",
	"Development Status :: 1 - Planning"
]

[project.scripts]
dotbackup = "dotbackup:main"

[project.urls]
Homepage = "https://github.com/jaxvanyang/dotbackup"
Source = "https://github.com/jaxvanyang/dotbackup"
Repository = "https://github.com/jaxvanyang/dotbackup.git"
Issues = "https://github.com/jaxvanyang/dotbackup/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/dotbackup.py"

[tool.hatch.build.targets.wheel]
packages = ["src/dotbackup.py"]
